🎨 [refactor]: Adjust Landing's structure
